Add Karma cog rewrite stub
